+++ CHANGES.md
@@ -1,3 +1,22 @@
+1.0.3 (2024-09-27)
+------------------
+
+- Added retry logic for all API endpoints
+- Added connect/read timeouts and an ability to ignose SSL certs
+- Noise Control Batch result API has its own retry parameter
+- All errors are standardised as {"status": "error", "message": "Reason"}
+
+
+1.0.2 (2023-12-22)
+------------------
+
+- Package updates
+
+1.0.1 (2023-12-22)
+------------------
+
+- Package updates
+
 1.0.0 (2023-12-22)
 ------------------